refactor(bridge): added context module
